requirements.txt
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
smtplib-secure>=1.0.0
joblib>=1.0.0
